+++ HISTORY.md
@@ -7,6 +7,7 @@ Legend:
 ### 0.2.2 (July 24th 2015)
 - [F] Import some internal modules correctly for Ubuntu/Linux node (case-sensitive for them).
 - [F] Nested springs work again.
+- [I]: Allow returning `null` from children function. #101
 
 ### 0.2.0 (July 22th 2015)
 - [B] `willLeave` returning `false` will now keep the key. Only `null` and `undefined` will serve as a signal to kill the disappeared key.